backend/requirements.txt
langchain
wikipedia
langchain-openai
langchain-community
langchain-anthropic
python-dotenv
pydantic
fastapi
uvicorn
requests
gunicorn==23.0.0
